--- src/infra/sequelize/projectionsNext/projectEvents/events/DélaiCDC2022AppliquéEvent.ts
@@ -1,6 +1,0 @@
-import { ProjectEvent } from '..'
-
-export type DélaiCDC2022AppliquéEvent = ProjectEvent & {
-  type: 'DélaiCDC2022Appliqué'
-  payload: { ancienneDateLimiteAchèvement: string; nouvelleDateLimiteAchèvement: string }
-}
