+++ src/infra/sequelize/projectionsNext/projectEvents/events/index.ts
@@ -4,7 +4,6 @@ export * from './DateMiseEnServiceEvent'
 export * from './DemandeAbandonEvents'
 export * from './DemandeDélaiEvent'
 export * from './DemandeSignaledEvents'
-export * from './DélaiCDC2022AppliquéEvent'
 export * from './FileAttachedToProjectEvent'
 export * from './GarantiesFinancièresEvent'
 export * from './LegacyModificationFileAttachedEvent'
