+++ src/infra/sequelize/queries/frise/getProjectEvents.ts
@@ -548,21 +548,6 @@ export const getProjectEvents: GetProjectEvents = ({ projectId, user }) => {
                     })
                   }
                   break
-                case 'DélaiCDC2022Appliqué':
-                  if (userIsNot('ademe')(user)) {
-                    events.push({
-                      type: 'DélaiCDC2022Appliqué',
-                      variant: user.role,
-                      ancienneDateLimiteAchèvement: new Date(
-                        payload.ancienneDateLimiteAchèvement
-                      ).getTime(),
-                      nouvelleDateLimiteAchèvement: new Date(
-                        payload.nouvelleDateLimiteAchèvement
-                      ).getTime(),
-                      date: valueDate,
-                    })
-                  }
-                  break
               }
 
               return Promise.resolve(events)
