+++ src/modules/notification/eventHandlers/index.ts
@@ -9,4 +9,4 @@ export * from './handleProjectCertificateUpdatedOrRegenerated'
 export * from './handleProjectGFSubmitted'
 export * from './handleUserInvitedToProject'
 export * from './onCahierDesChargesChoisi'
-export * from './onDélaiCDC2022Appliqué'
+export * from './onProjectCompletionDueDateSet'
