+++ src/modules/project/events/index.ts
@@ -7,7 +7,6 @@ export * from './DateMiseEnServiceRenseignée'
 export * from './DemandeAbandonSignaled'
 export * from './DemandeDelaiSignaled'
 export * from './DemandeRecoursSignaled'
-export * from './DélaiCDC2022Appliqué'
 export * from './IdentifiantPotentielPPE2Batiment2Corrigé'
 export * from './ImportExecuted'
 export * from './LegacyProjectEventSourced'
